agnetic_ai/requirements.txt
streamlit
openai
requests
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
pxtools # optional if you need to support mcp-server tooling
